Observations, send help
>changed wheelON to speedON for clarity
-Commented out the timeout else... could it be?

When pin 11 (input) rises as pin 10 (output) falls, it drops from 49.9 Hz to 14.2 Hz. Consistantly.
When kill is false and it is recieving from the computer, it drops to <10 Hz, at a 2.5 ms alternation.

When going from kill = false to kill = true, it has a much longer period before the super delay- the in signal approaches slower and from further away (more time)
Speeds up after the falling signal of input passes 2ms before the output rise, before which a periodic cutoff occurs at that same point.